Rework deg, sec and min methods for vposition
Precision problems with conversion back and forth between degrees and radians
caused incorrect output for min and sec values, for example 30m00s became
29m60s. Same value but not correctly displayed. Lowering the precision from
double to float when computing number of minutes solved the problem. Computing
the seconds relies on the min-methods and the "precision loss technique"
therefore propagates.

Still not completely solved, i.e. the problem behind this and what better
technique to use is still unknown.